renamed script files, moved all potency methods to separate file
