+++ hikari/dataframes/hkl.py
@@ -432,6 +432,17 @@ class HklFrame(BaseFrame):
         """
         return 2 / self.la
 
+    def _in_dacs(self, opening_angle, vectors):
+        oa = np.deg2rad(opening_angle)                # opening angle in radians
+        xyz = self.table.loc[:, ('x', 'y', 'z')].to_numpy()
+        r = self.table.loc[:, 'r'].to_numpy()
+        v = np.array(vectors)
+        v = (v.T / lin.norm(v, axis=1)).T              # normalise vectors v
+        m1 = np.matmul(v, xyz.T)                       # dist from dac plane "p"
+        phi = np.abs(np.arcsin((m1 / r).clip(-1, 1)))  # angle <(plane p, v)
+        lim = self.r_lim * np.sin(oa - phi)            # True if in dac
+        return r[None, :] < lim
+
     def dac_trim(self, opening_angle=35.0, vector=None):
         """
         Remove reflections outside the opening_angle DAC-accessible volume.
@@ -459,17 +470,6 @@ class HklFrame(BaseFrame):
         self.table = self.table[in_dac]
         self.table.reset_index(drop=True, inplace=True)
 
-    def _in_dacs(self, opening_angle, vectors):
-        oa = np.deg2rad(opening_angle)  # opening angle in radians
-        xyz = self.table.loc[:, ('x', 'y', 'z')].to_numpy()
-        r = self.table.loc[:, 'r'].to_numpy()
-        v = np.array(vectors)
-        v = (v.T / lin.norm(v, axis=1)).T  # normalise vectors v
-        m1 = np.matmul(v, xyz.T)  # dist from dac plane "p"
-        phi = np.abs(np.arcsin((m1 / r).clip(-1, 1)))  # angle <(plane p, v)
-        lim = self.r_lim * np.sin(oa - phi)  # True if in dac
-        return r[None, :] < lim
-
     def dacs_count(self, opening_angle=35.0, vectors=np.array((1, 0, 0))):
         """
         Count unique dac-accessible reflections for n crystals placed such that
@@ -480,7 +480,7 @@ class HklFrame(BaseFrame):
         :param vectors: Array containing rotational axes of available DAC-discs.
         :type vectors: np.array
         :return: Array with numbers of unique reflns in DAC-accessible region.
-        :rtype: int
+        :rtype: np.array
         """
         memory_estimate = 26 * len(self) * len(vectors)  # estimate memory use
         cycles_needed = -(memory_estimate // -hikari.MEMORY_SIZE)  # and split
@@ -488,10 +488,10 @@ class HklFrame(BaseFrame):
             vectors_split = np.array_split(vectors, cycles_needed)
             return np.hstack([self.dacs_count(opening_angle, vectors=v)
                               for v in vectors_split])
-
-        in_dac = self._in_dacs(opening_angle, vectors)
-        return np.array([self.table.loc[in_dac[n, :], 'equiv'].nunique()
-                         for n in range(vectors.shape[0])])
+        else:
+            in_dac = self._in_dacs(opening_angle, vectors)
+            return np.array([self.table.loc[in_dac[n, :], 'equiv'].nunique()
+                             for n in range(vectors.shape[0])])
 
     def copy(self):
         """
